push batch streaming into ScanRecorder; pin pre_buffer in point lookup
- ScanRecorder gains an abstract results_batches() static method;
  FileRecorder implements it (parquet iteration + value_type
  uniformity pre-pass moved from _scanresults.py into the recorder)
- scan_results_batches() is now a thin wrapper that dispatches via
  scan_recorder_type_for_location and applies the same per-batch
  transformations as scan_results_df
- _point_lookup pins pre_buffer=True explicitly (default changed in
  pyarrow 25) since it reads single row groups where coalescing is a
  free win
